net: phy: smsc: shorten energy detect off time in lan87xx_read_status() [STREAMEIGHT-2724]
This essentially reverts the following commit:
776829d net: phy: workaround for buggy cable detection by LAN8700 after cable plugging

Main reason being is the power consumption, which is considerably higher
when disabling energy detect mode for 640 ms instead of 64 ms. Also, since
the kernel regularly calls read_status() it should be not an issue if link
detection fails once (oe even more times).

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>
[rebased for 6.1]
[rebased for 6.6]
Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>