ASoC: fsl_sai: add handling of SND_SOC_DAIFMT_CONT
In many cases we want to have the audio clocks running all the time to
minimise audible click artifacts on playback start/stop. However the
fsl_sai driver will disable the audio clocks when no playback is active.

The ASoC core has a special continuous clock flag for that reason
`SND_SOC_DAIFMT_CONT`.

This change adds the handling of this flag, so the card driver is able to
specify this flag and the clocks will keep running even after a playback
has stopped.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>
[rebased on top of 5.4.47_2.0.0]
Signed-off-by: Quentin Schulz <quentin.schulz@streamunlimited.com>
[rebased on top of 5.10]
[rebased on top of 6.1]
[rebased on top of 6.6]
[rebased on top of 6.18]
Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>