drivers: soc: imx: add simple-busfreq-imx8mm driver
The original NXP driver for this purpose `busfreq-imx8mq.c` is overly complicated
since it tries to have all the switching logic inside the kernel. Additionally
the switching logic is not very clear inside the NXP driver and it has many
conditionals depending on the platform e.g. i.MX8MQ vs i.MX8MM, etc.

So instead of cleaning up this mess of a driver, we just write a simple
alternative which is only for the i.MX8MM and has no switching logic inside
the driver. Instead it exports a sysfs file
`/sys/devices/platform/simple-busfreq/set_high` where the userspace decides
when to enter the low bus frequency mode.

The targeted frequencies are:
* HIGH mode:
  * DDR4:    2400 MT/s
  * NOC:     750 MHz
  * AHB:     133 MHz
  * AXI:     333 MHz
* LOW mode:
  * DDR4:    400 MT/s
  * NOC:     150 MHz
  * AHB:     22 MHz
  * AXI:     25 MHz

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>
[replaced ARCH_FSL_IMX8MM with ARCH_MXC && ARM64 in depends on]
[replaced include fsl_sip.h with FSL_SIP_DDR_DVFS constant definition]
[modified comments and commit log to give the DDR4 freqs]
[replaced device_create_file with sysfs_create_file to match 5.4 driver]
[reworked based on newer busfreq-imx8mq.c which: ]
[ - does not register irq handler for WFE but let ATF take care of it,]
[ - requests ATF for all supported freqs by the DRAM chip,]
[ - sends the *index* in the freq table of the requested frequency to ATF for DDR DVFS,]
[ - makes the 5.4 kernel depend on ATF >= imx_5.4.3_2.0.0,]
Signed-off-by: Quentin Schulz <quentin.schulz@streamunlimited.com>
[rebase to 6.6]
Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>