S195xDTS: document how to enable uart4 in Linux
UART4 IP domain access is specified in RDC_PAP70 (0x303D0518).

It is set to deny RW rights to domain 0 RW which is where all the
masters (c.f. table 3-2 in i.MX8MM Reference Manual) are by default.

Domain 1 is the Cortex M4's. So the UART4 is accessible only by the M4.

One can change that so that the domain 0 (hence cortex a53 running
Linux) can access it too.

Signed-off-by: Quentin Schulz <quentin.schulz@streamunlimited.com>