S195xDTS: add support for StreamKit 1955 devkit [STREAMEIGHT-4403]
Compared to the current state of support for StreamKit3D, StreamKit1955
now have an AK5704 for microphones (on the 4mic breakaway PCB) which
does the PDM to PCM translation instead of the imx8 doing it directly.

The AK5704 lanes are connected to SAI5 and the output of AK4458 is now
connected to SAI1.

The mic mute status is not read from a GPIO anymore but handled by the
MCU directly.

Signed-off-by: Quentin Schulz <quentin.schulz@streamunlimited.com>
[replaced ak4458,pdn-gpio with reset-gpios]
[replaced IMX8MM_CLK_SAI1_SRC and IMX8MM_CLK_SAI1_DIV with IMX8MM_CLK_SAI1 and removed rate]
[replaced IMX8MM_CLK_SAI5_SRC and IMX8MM_CLK_SAI5_DIV with IMX8MM_CLK_SAI5 and removed rate]
[moved dai-tdm-slot-num and dai-tdm-slot-width to dai subnodes]
[replaced nand version with emmc]
Signed-off-by: Quentin Schulz <quentin.schulz@streamunlimited.com>