ASoC: codecs: add support for AK5704 [STREAMEIGHT-4403]
The AK5704 IC is taking PDM input from the microphones and outputting
the data in I2S format.

This driver supports only a minimal subset of the featureset supported
by this IC. The compatible rates are limited by the microphones
populated on the 4mic board available for the StreamKit1955 (namely 8k
and >48k sampling rates are not supported).

The IC supports only powering on selected microphone inputs which is a
nice addition to probably gain some power consumption if not recording
on 4 channels.

The driver expects only digital input microphones (PDM) with two
I2S physical lines with two audio channels per physical line.

The driver expects to create the internal PLL from the BCLK pin provided
by the SoC DAI and also expects the LRCLK to be provided by the SoC DAI.

For now, the only supported format is S32_LE, though the other one
supported for digital input mics by the IC is S16_LE.

Signed-off-by: Quentin Schulz <quentin.schulz@streamunlimited.com>
[replaced snd_soc_codec with snd_soc_component]
[replaced snd_soc_write with snd_soc_component_write]
[replaced snd_soc_codec_driver with snd_soc_component_driver]
[replaced snd_soc_register_codec with devm_snd_soc_register_component]
[removed useless remove function as devm take care of unregistering the component]
[added comment on the need for pcm constraints and SNDRV_PCM_RATE_KNOT]
[added comment on where the pll table for given rates is found]
Signed-off-by: Quentin Schulz <quentin.schulz@streamunlimited.com>
[rebase to 6.6]
Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>