S195xDTS: fix incorrect overrides of assigned-clock-* of clk controller node
In original commits on the 4.14 kernel, assigned-clocks weren't set in
the fsl-imx8mm.dtsi clk controller node, therefore it had to be defined
to set the expected clock rates.

In 5.4 kernel, imx8mm.dtsi already has the correct assigned-clocks set
(and even more), therefore there shouldn't be an override in s1955
device trees.

The point of the clk override was to double the rate of
IMX8MM_AUDIO_PLL1 and IMX8MM_AUDIO_PLL2 clocks, therefore one just needs
to copy the original assigned-clock-rates and double the appropriate
rates in the array.

Fixes: 8fa684a03580 "S195xDTS: add support for StreamKit 1955 devkit [STREAMEIGHT-4403]"
Fixes: aaa7276888c8 "S195xDTS: add fsl-imx8mm-sue-stream195x-l0-kit.dts file"

Reported-by: Martin Geier <martin.geier@streamunlimited.com>
Signed-off-by: Quentin Schulz <quentin.schulz@streamunlimited.com>