S195xDTS: add support for new 4-microphone board based on TI PCMD3180 [STREAMEIGHT-4547]
The PCMD3180 is a TLV320adc3140 stripped of most of its ADC features,
but the driver is compatible, though showing inapplicable mixer elements
in ALSA.

The PCMD3180 is configured to do multi-slotting (channel 1/3 and 2/4 are
output to respectively left slot and right slot) and multi-outputting
(channel 1 and 2 to SDOUT, 3 and 4 to SDOUT2). It provides PDM clock to
the PDM microphones via GPO pins and channels 1 to 4 are enabled by
default at boot. GPIO1 is used as SDOUT2.
PDM microphones 1 and 2 are routed to PDMIN1 and PDM microphones 3
and 4 to PDMIN3.

Signed-off-by: Quentin Schulz <quentin.schulz@streamunlimited.com>