ASoC: stream195x: support dummy codecs
There is currently no support for dummy codecs in the stream195x sound
card driver (only for DPCM it seems so, which is not used on Stream1955
yet). Thus if a dummy codec is wanted, the sound card device will fail
to probe.

This adds support for dummy codecs in stream195x sound card driver by
trying first to find a valid phandle in sound-dai property (current,
default behavior), if it is invalid or the property is not found, it'll
look for a "use-dummy-codec" boolean property instead. If it is present,
then the snd-soc-dummy device registered by the ASoC core in
sound/soc/soc-utils.c.

Fixes: 98b1fa8d4524 "ASoC: stream195x: add Stream195x sound card"

Reported-by: Martin Pietryka <martin.pietryka@streamunlimited.com>
Tested-by: Martin Pietryka <martin.pietryka@streamunlimited.com>
Signed-off-by: Quentin Schulz <quentin.schulz@streamunlimited.com>
[rebase to 6.6]
Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>