dt-bindings: sound: stream195x: fix dummy codec explanation and example
The dummy codec support as explained currently is actually not working.

A change was made in the sound card driver to check for use-dummy-codec
property when sound-dai does not exist or points to an invalid device.

Therefore, having a codec DAI node with only use-dummy-codec property in
it is now the way to specify dummy codecs.

Fixes: 98b1fa8d4524 "ASoC: stream195x: add Stream195x sound card"

Reported-by: Martin Pietryka <martin.pietryka@streamunlimited.com>
Signed-off-by: Quentin Schulz <quentin.schulz@streamunlimited.com>