ASoC: codecs: ak5704: fail codec probe if device absent
The i2c driver probe function always succeeds even if the device is
absent from the bus.

It is up to the codec probe to fail if the device isn't there, which can
be done by checking the return code of snd_soc_component_write and
return it if it's not zero.

This fixes "ghost" ak5704 device in userspace.

Fixes: af519b1d7225 "ASoC: codecs: add support for AK5704 [STREAMEIGHT-4403]"

Signed-off-by: Quentin Schulz <quentin.schulz@streamunlimited.com>