S195xDTS: put PDM-I2S devices in their own sound card
Since the PDM-I2S devices are on the 4mic board which is an
daughterboard/external board/pluggable board/whatever you name it, and
can be absent (in valid configurations, e.g. Google Cast builds), the
sound card will not probe currently.

This results in a valid configuration not being usable.

By having an additional sound card per external daughterboard, it is now
possible for a device to not be present and fail only its own sound
card, while the rest works as well as before.

Signed-off-by: Quentin Schulz <quentin.schulz@streamunlimited.com>