S195xDTS: do not configure RX path for SAI1
It's important for DAB/AUX_IN that RX path of SAI1 is not trying to be
the bitclock master, since it would fight with the ADC clock. The TX
path, however, needs to be configured as master, as this is our main
audio output path.

Simply not initializing the RX pins should be enough, since they are
only used by the Audio Framework, and not needed by Linux.

Signed-off-by: Andriy Danylovskyy <andriy.danylovskyy@streamunlimited.com>