ALSA: pcm: fix audio glitch
A short portion of audio from previous session was played at the start of
the next session.

Previously the code which is generating the silence (and thus cleaning the
whole audio buffer) was called from within the snd_pcm_post_start()
function. This was incorrect as the playback could have already been
running at that time (as it is started in snd_pcm_do_start()). The buffer
cleanup was therefore moved up to snd_pcm_do_start() before then moment
when the trigger() is called which guarantees that no relicts of old audio
are being played.

Signed-off-by: Stanislav Ruzani <stanislav.ruzani@streamunlimited.com>
[follow maximum line length]
[do not use brackets for one line statements]
Signed-off-by: Radek Dostál <radek.dostal@streamunlimited.com>
[rebased for 6.1]
Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>
[rebase to 6.6]
Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>