S1832Config: enable iptables support [STREAMEIGHT-5205]
These configs form a minimal set required so that iptables actually work.
We also want 'conntrack' support, if only for allowing outgoing connections.
It was decided to keep
 * CONFIG_NETFILTER_NETLINK_LOG and CONFIG_NETFILTER_XT_TARGET_LOG
 * CONFIG_NETFILTER_XT_MATCH_COMMENT
 * CONFIG_NETFILTER_XT_MATCH_MULTIPORT
 * CONFIG_IP_NF_TARGET_REJECT
as they are generally useful.

There is just one table - filter, because nat/mangle are not needed. It's
more than enough for out need, which is "filtering of incoming connections
based on TCP/UDP port".

Signed-off-by: Maros Medvec <maros.medvec@streamunlimited.com>
[running savedefconfig again removed CONFIG_IPV6_SIT]
Signed-off-by: Radek Dostál <radek.dostal@streamunlimited.com>
[rebased for 6.1]
[rebased for 6.6]
Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>