pincrtl: meson: set gpio base addr based on dt
The gpio sysfs interface is deprecated but still used heavily in our
codebase and we rely on the gpio addresses to be the same every time the
kernel boots. This is not supported by default, because the sysfs
interface is not prepared to assign gpios in a predictable manner.
However for backwards compatibility we need to handle this usecase.

Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>
[rebase for 6.6]
Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>