mmc: sdhci-esdhc-imx: allow to reset sdhci driver [STREAMEIGHT-5564]
Implement a way to reset the sdhci device by unbinding and re-binding
the driver. This is the last resort recovery for cases like firmware
crashes, that are hard to handle within the driver logic.

This will make the broadcom wlan device disappear and re-appear in the
system, thus triggering re-configuration on the udev/app level.

It can be called in two ways:
1. exported function sdhci_esdhc_rebind_driver
2. sysfs entry /path/to/mmc/driver/rebind:
echo -n "30b40000.mmc" > /sys/devices/platform/soc@0/subsystem/devices/30b40000.mmc/driver/rebind

Signed-off-by: Andriy Danylovskyy <andriy.danylovskyy@streamunlimited.com>