clk: meson: fix HIFI-PLL fractional support
The upstream driver does not handle fractional parts that are negative.
However the axg HIFI-PLL-DCO has a 15 bit fractional register where bit 14
represents the sing of the fraction. This feature is used by the drift
compensator. The upstream driver ignores this for some reason and only
uses 13 bits instead of the available 15

This commit also adds the HIFI-PLL tables which have been
calculated using script done by Emil Petersky
(STREAMEIGHT-3241: [1832] enable HiFi PLL).

Lastly we set the downstream divider of the HIFI-PLL-DCO to 4 always,
by making the HIFI-PLL divider read-only.
This slight optimization helps the clk_fw to find the correct parameters
for the MCLK divider when propagating the rate change to the parent.
This way we don't have to call set_rate on the HIFI-PLL-DCO
in axg_tdm_iface_set_sysclk(), instead it can just conviniently set
the MCLK and the rest will be handled by the HIFI-PLL-DCO and the
MCLK divider; matching the same behaviour we had on the old 4.19 vendor kernel.

Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>
[rebased for 6.18]
Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>