clk: meson: axg-audio: enable SET_RATE_PARENT only for mst_mclk and spdifin_clk
We run the TDM and PDM audio clocks from the HIFI_PLL. SPDIF in runs from mpll.
When the PDM mic sets the rate, it should not recalculate the HIFI_PLL because
that is set by the drift-compensator by one of the TDM channels.
This commit introduces a two construct, called AUD_MST_MCLK_MUX_SET_PARENT
and AUD_MST_SYS_MUX_SET_PARENT which are used to allow the TDM and
SPDIF to set the parent rate.
This way when the PDM driver sets the rate for the microphone the HIFI_PLL
is not disturbed.

Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>
[squashed: clk: meson: axg-audio: enable SET_RATE_PARENT for spdifin_clk [STREAMEIGHT-6756]]
Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>