mmc: meson-gx: enable CFG_AUTO_CLK in SD_EMMC_CFG
This bit was also enabled in the vendor kernel. The meaning of this bit is
described in the documentation as follows:

"""
cfg_auto_clk: SD/eMMC Clock Control:

1: when BUS is idle and no descriptor is available, automatically turn off
 clock, to save power.
0: whenever read data core from clock SD/eMMC is on the SD/eMMC clock is ON,
 it is still on/off
"""

Which means this bit turns off the clk when there is no data, to save power.

According to Mehmed Mulaomerovic from hw dept, setting this bit will also
decrease EM emission.

The 4.9 and 5.4 vendor driver sets this bit in the card_busy() callback
so we copy the same behavior.

Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>
[squashed: mmc: meson-gx: disable CFG_AUTO_CLK during reset [STREAMEIGHT-6137]]
Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>