S195xDTS: enable pull-down resistor on MDIO data line [STREAMEIGHT-5769]
If the MDIO data line is not pulled low and no PHY is connected (such as on
S1955 evaluation board), MDIO reads will return 0xffff for all registers.
This makes the PHY subsystem think that there is a generic phy and that
the link is up, which results in swupdate trying to fetch IP address
forever.

I'm not 100% sure why this was not needed on the old kernel, but talking to
HW department it's an acceptable fix for when there is no PHY. The correct
solution would be to always set correct "carrierboard" in the const
partition, but that may lead to way more problems than this simple fix.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>