usb: ci: core: fix vbus over-current workque in suspend to ram [STREAMEIGHT-5647]
When suspending to ram ci_controller_resume() is called, which schedules
check_vbus_work. Unfortunately this work reschedules itself when the
callback is done, so the workqueue got stuck because flush_workqueue()
was called in ci_suspend(). This resulted in the infinite schedule of check_vbus_work
which prevented the workqueue from flushing, which led to a hang when suspending to ram.

To fix this, the flush_workqueue() call was dropped because the workqueue
is freezable so any work scheduled will continue on wakeup.

To support this, the logic in imx8_check_vbus_overcurrent() had to be modified.
Since ci_runtime_suspend() returns -EBUSY when an overcurrent is detected,
imx8_check_vbus_overcurrent() can assume that the overcurrent bit can be read
when device is experiencing an overcurrent. So after reenabling the regulator,
instead of rescheduling the task and potentionally letting the device enter
low power mode, the function now checks after some delay if the OC condition still
happens. This is now not a rescheduled task, but happens synchronously.
If the OC is fixed, it reschedules the task as usual and clears the flag.
If the OC is still happening after the small delay, the task is resceduled.

The synchronous delay was added, because when OC is active, the device
is not in low power mode so the OC condition bit can be read.
Rescheduling could put the device to runtime_suspend() which would block
the reading of this bit, because the OC flag is cleared at that point.

Fixes: fe485d0162f1 ("usb: ci: core: fix vbus over-current protection workqueue")

Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>
[rebased to 6.6]
Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>