ASoC: stream195x: add sysfs file to control suspending of card components [STREAMEIGHT-3572]
This patch increments the use count (using pm_runtime_get_snc()) on
late_probe() once so any components (Codec, DAC, SAI, etc.) will not be
automatically suspended on playback stop. (Previously we hacked around this
in the respective component drivers by basically ignoring/doing nothing in
the suspend() callbacks.

A new sysfs file called "ignore_suspend" was added to control this
behavior, the default is 1 so that card components are not suspended
when playback stops. Writing 0 to this will will decrement the reference
count of the card components again allowing for them to be suspended
once audio playback stops.

This is a reasonably generic solution, the components will be resumed and
suspended as needed and the behavior can be controlled from the userspace
using a single sysfs file.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>
[rebase to 6.18]
Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>