ASoC: fsl_sai: switch pinmux in runtime/suspend callbacks [STREAMEIGHT-3572]
This allows us to switch between pinmuxes when suspending and resuming the
driver.

Also the previous useless handling of sai->suspended was dropped, the
resume/suspend callbacks will correctly enable/disable the corresponding
clocks if needed. Since clocks are refcounted this is fine for the case
where we want to have continuous clocks as in that case we already
incremented the refcount of all MCLKs once in fsl_sai_set_dai_fmt().

NOTE: The new pinmux entry must be called "suspend", if it is called
"sleep" the "soc-pcm" core will automatically select the "sleep" pinmux
when the PCM device is closed, which is not what we want.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>