watchdog: meson: disable keepalive messages sent by kernel after boot
Clear WDOG_HW_RUNNING bit added in commit b91a395d7e06
("watchdog: meson: keep running if already active") because it
would send keepalive to watchdog while kernel is running, but this is not
the behavior we want. Keepalive should only be coming from userspace while
watchdog is active.

Also fix the timeout which is set to the max value, use DEFAULT_TIMEOUT
once the watchdog is started.

Fixes: b91a395d7e06 ("watchdog: meson: enable watchdog in probe")

Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>