mtd: nand: raw: amlogic: hide info pages from bootloader partition
On the bootloader partition (mtd0) every block has 2 info pages which
cannot be written by userspace, only the driver has access to these and
it keeps it consistent. Probably has to be kept like this for the rom
bootloader.

These blocks could be read back by userspace which caused issues when
trying to verify if the bootloader was flashed correctly.

On the old vendor kernel these blocks were hidden. Copy the behavior of
the old vendor kernel and shift every page so that these info pages are
hidden. This page shift also means the last page is now not available, so
prevent read failures by returning -ENODATA.

Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>