S1832DTS: disable hwrom_reserved and free ~10MB of memory
hwrom_reserved is a conservative way to handle whether BL32 was loaded or
not by U-boot [0]. U-boot should delete this region from the DTS when BL32
is not loaded and should update the DTS with proper sizes when it is loaded.

Unfortunately our old vendor u-boot does not handle hwrom_reserved at all
so this memory is reserved on top of BL32 even when B32 sizes are properly
set by modifying secos_reserved and secmon_reserved from U-boot.

[0] https://libera.irclog.whitequark.org/linux-amlogic/2023-03-28

Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>