drivers: regulator: apx15060: fix shutdown issue [STREAMEIGHT-6171]
When `shutdown -h now` was called the system did not halt.

This happened because the I2C adapter was in suspend. Unfortunately devm_register_sys_off_handler
requires an atomic xfer handler but that can't prepare the clk because
that can only be done in non atomic context. So the I2C adapter has to be
waken up first so the clk is prepared properly.

Fixes: f4810fa6fc9e ("drivers: regulator: apx15060: add power off handling [STREAMEIGHT-4277]")

Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>
[adapted to kernel 6.1]
Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>
[squashed fix c2286d96f3 "drivers: regulator: apx15060: fix shutdown issue [STREAMEIGHT-6768]"]
Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>