S195xCONFIG: disable firmware fallback mechanism [STREAMEIGHT-6255][STREAMEIGHT-6317]
When the kernel tries to load firmware, if it is unable to find it in
the /lib/firmware/ directory, it may attempt to load it via a "fallback
mechanism" based on uevents.

The CONFIG_FW_LOADER_USER_HELPER option enables this fallback
mechanism to be used. However if udev is not active when this fallback
mechanism is used, the request times out for a minute. This caused
excessively long boot times for the factory testing image, where the wifi
driver is loaded directly from the init script and not via udev.

Furthermore, this fallback mechanism has been marked as deprecated, and
udev dropped support for it in 2014 [1]. Yocto Kirkstone is still
re-enabling the support by reverting the delete patches [2], but Yocto
Langdale no longer does that [3].

[1] https://github.com/eudev-project/eudev/commit/3b717594600fa717cdf9bcfd0c7c1b703b245482
[2] https://github.com/openembedded/openembedded-core/commit/d6b139ef1b52ee4842f8706c1b8b950cc50a3d54
[3] https://github.com/openembedded/openembedded-core/commit/31005f24d302082c98bdd33c1fc0fb3db81c74e4

Signed-off-by: Malik Tuwebti <malik.tuwebti@streamunlimited.com>
[run savedefconfig]
[disable CONFIG_FW_LOADER_USER_HELPER too]
[provide yocto references]
Signed-off-by: Radek Dostál <radek.dostal@streamunlimited.com>