rtkdrivers: cfg80211_fullmac: map KM4 RAM into driver and use it
With the previous changes the Linux memory starts at 0x60800000 (8 MiB)
compared to previously starting at 0x60300000 (3 MiB).

The Realtek WiFi uses a shared memory region somewhere at the beginning
of the RAM (0x60000000) to exchange data with the M4.

Before moving the start of the Linux RAM up, the whole RAM was covered
by the static virt_to_phys()/phys_to_virt() mapping, this was probably
by coincidence.

After moving the start of the Linux RAM up to 0x60800000, the kernel
will no longer do this static mapping from the beginning of the RAM.

For that reason we need to correctly map the used region at the beginning
of the RAM using devm_ioremap_nocache() and create a new helper function
km4_phys_to_virt() to do the calculation for us.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>