rtkdrivers: pinctrl-amebad2: use postcore_initcall instead of arch_initcall
The postcore_initcall make sure that the pinctrl driver is loaded and
registered before the gpio driver.

Otherwise if we have a pinctrl on a node, the probing of the GPIO driver
would have been deferred which then created issues with other drivers
relying on the GPIO driver and not handling the defer cleanly, e.g. SPI
MIPI.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>