arm: kernel: setup: set system_serial from generated unqiue ID in the KM4
The Ameba chip does not have a built-in unqiue ID, so instead the KM4
bootloader will derive an unique ID from the HUK and place it in a
non-secure SRAM retention region with the address of 0x23020170.

This change allows the userspace to fetch the unique ID from the
`/proc/cpuinfo` file.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>