ASoC: realtek: sport: limit minimum channels to 2 and max sample-rate to 96 kHz
Mono playback currently does not work properly, this is probably related
to some wrong setup of the DMA. Since this is not a requirement for us,
we can just enforce at least 2 channels.

In a similar vein, there are also issues and glitches with 192 kHz
playback, thus limit the maximum sample-rate to 96 kHz.

If we really need support for those things we can do a more in-depth
investigation later on.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>