S195xCONFIG: build IMX_SDMA as kernel module
By default IMX_SDMA driver probes during very early stages of boot. At
that time it can not load /lib/firmware/imx/sdma/sdma-imx7d.bin which is
located in rootfs and therefore it has to rely on userspace fallback
mechanism [1].

This udev is not present in our factory images and additionally it has
been deprecated [2]. Yocto Kirkstone is still re-enabling the support by
reverting the remove patches [3], but next Yocto version no longer does
that [4].

Making IMX_SDMA a module ensures, that it will only be probed at the time
of rootfs being present. This should not be a big deal as before the SDMA
was also only fully operational when it was able to load the firmware from
the rootfs.

[1] https://www.kernel.org/doc/html/v6.1/driver-api/firmware/direct-fs-lookup.html
[2] https://github.com/eudev-project/eudev/commit/3b717594600fa717cdf9bcfd0c7c1b703b245482
[3] https://github.com/openembedded/openembedded-core/commit/d6b139ef1b52ee4842f8706c1b8b950cc50a3d54
[4] https://github.com/openembedded/openembedded-core/commit/31005f24d302082c98bdd33c1fc0fb3db81c74e4

Signed-off-by: Radek Dostál <radek.dostal@streamunlimited.com>