rtkdrivers: panel: ameba_panel: use module_platform_driver() macro
There is no reason to use postcore_initcall() which will try to call the
init function early in the boot and creating issues with missing GPIO and
pinctrl drivers. Instead just use the normal module_platform_driver() macro
which 1. is cleaner and 2. moves the actual probe of the panel driver to
a later stage.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>