net: bluetooth: smp: use old ecdh crypto names
Realtek decided to backport this file or some parts of it to their kernel,
we cannot see their history or details. But in the process of back-porting,
they also backported the new names for the ECDH crypto primitives. However
in this version of the Linux kernel these new crypto primitive names are
not yet available.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>