ASoC: meson: warn about SPDIF rate mismatch [STREAMEIGHT-6411]
SPDIF can't set the parent clk because commit d0cd10e ("clk: meson:
axg-audio: enable SET_RATE_PARENT only for mst_mclk") only allows HIFI_PLL
to be set by TDM interface. If the TDM configures the PLL for a base rate
that is not compatible with the requested rate for the SPDIF playback the
rates can be mismatched. This results in CLK drift over time.

We can't allow the SPDIF to set the parent CLK because that would ruin the
settings set by the TDM interface. Instead we warn the user about the
potential misconfiguration.

To do this we check if the requested rate is within 500 ppm of the actual
rate. This is necessary because of two things:
1) HIFI_PLL does not have the precision to set the rate exactly
2) Drift compensator might alter the base PLL freq within 500 ppm range

Sometimes the CLK system can find a suitable divisor even if the parent
base rate is not aligned with the requested rate. In that case we can't
warn about the mismatch unfortunately.

We could work around this by going to the parent CLK and checking that
instead but that would be a solution which would cause a lot of hard-coding
in this driver. So instead we warn the user with best effort.

Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>