ASoC: fsl_sai: also disable clocks internally when entering suspend [STREAMEIGHT-3572]
Previously when entering suspend, only the pinmux was changed to disable
the clocks. However internally the MCLK was still active. This patch
changes this behavior so that when actually entering the suspend state, also
internally the clocks will be disabled. This probably does not gain us much
in terms of power-savings, but it's cleaner to do it.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>