rtkdrivers: spi: do not unconditionally re-enable the DMA support
In the device-tree the rtk,spi-dma-en property is set to disable the DMA,
however the Realtek code in the recent release added some functionality
to the callback that enables the DMA suddenly when the transfer length
exceeds RTK_SPI_FIFO_ALL. This does not make sense and it also crashes
the kernel when the DMA is actually used. For that reason honor the DMA
setting from the device-tree for now. Ideally we should dig into this and
figure out what the intentions of Realtek were to add this behavior.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>