HACK: ASoC: realtek: machine: add `sue,disable-analog-links` property [STREAMEIGHT-6321]
Some pins of the Realtek SoC are shared between an internal analog audio
DAC and the digital port. Unfortunately the current pinmux driver has no
way of switching those pins between analog and digital operation. What
works however is if the analog DAC/Codec link is disabled, in that case the
pins will not be reconfigured for analog operation. We cannot fully disable
the audio driver, as we need the I2S output to work, so instead we add this
property to disable only the analog links of the sound card.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>