rtkdrivers: cfg80211_fullmac: add shared header files from rtl8730e-linux-sdk-3.0_r4-beta release
With the last sdk-3.0_r4-beta release Realtek started sharing the
headers between the Linux driver and the firmware. Since we are building
both separately we need to copy and commit them to be able to build the
Linux driver.

The respective header files were taken from the
`sources/firmware/component/wifi/driver/include/` directory in the
rtl8730e-linux-sdk-3.0_r4-beta release.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>