rtkdrivers: dma: fix rtk_dma_free_lli() race condition
It could happen that rtk_dma_free_lli() was called from the PCM/dmaengine
system through the rtk_dma_terminate_all()->rtk_dma_free_txd() calls. At
the same time a call to rtk_dma_free_lli() could occur from the DMA
interrupt.

This led to a race condition where the list could have been in the middle
of being modified as the interrupt came in. To prevent that make sure that
the interrupts are disabled and the vc.lock spinlock is acquired in
rtk_dma_terminate_all() when calling rtk_dma_free_txd().

This commit locks and disables interrupts for the whole function as
potentially other fields are modified as well, this is not an issue as
rtk_dma_terminate_all() is not called often. Also other dmaengine callbacks
(e.g. rtk_dma_pause()) similarly acquire the lock and keep interrupts
disabled for the whole duration of the respective call.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>