ASoC: codecs: ak4458: only call mdelay() in ak4458_set_dai_mute() if register has changed [STREAMEIGHT-6709]
SNDRV_PCM_IOCTL_PREPARE is called multiple times during snd_pcm_hw_params() and
snd_pcm_prepare() in alsa-lib. The mute function is called inside prepare and
it results in huge delays because regardless of the current mute state,
the mdelay is called.

snd_soc_component_update_bits() returns 1 if the state of the register is changed
and 0 when it stays the same. So mdelay should only be needed if the mute state
has changed in the AMP.

Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>