ASoC: meson: spdifin: add rate constraint when rate is detected [STREAMEIGHT-6669]
Don't allow the S/PDIF to be opened at any arbitrary rate if a rate is
currently detected. For example opening the S/PDIF with 48KHz while on the
input we see 88.2KHz would mess up the audio data.

So limit the rate to the actual rate we see on the source.

Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>
[polished commit message and coding style]
Signed-off-by: Radek Dostál <radek.dostal@streamunlimited.com>