rtkdrivers: phy-rtk-usb: use calibration data from previous release
Realtek reworked the USB calibration and it does not really work on our
boards anymore. There is a ticket at Realtek pending for this to provide
more information/instructions for calibration.

However, for now to make USB work again, we will use the calibration table
from the previous release. It was not a table previously, but just code
however this table was "reconstructed" from the previous code flow and works
on our boards.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>