rtkdrivers: re-add realtek-otp-core driver
We will need this driver for now to be able to write the secure fuses from
the non-secure Linux world through the KM4 using the IPC mechanism.

The driver is taken from the `rtl8730e-linux-sdk-3.0_r3-beta` release
but the compatible was changed from "realtek,amebad2-otp" to
"realtek,amebad2-otp-core" and the `driver.name` was changed from
"rtk-otp" to "rtk-otp-core" to not clash with the currently existing
driver.

The currently existing driver will be modified to use the
`rtk_otp_process()` function whenever we access the physical secure
fuses.

Similar for the `IPC_USER_*` enum and the `struct rtk_otp`, these were
moved from the `realtek-otp-core.h` header to the driver. This header
was actually obsolete and should have been removed, but now that we
re-added the otp-core driver we need this header for now anyway.

Additionally the Realtek copyright header was added to the driver as
well.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>