rtkdrivers: otp: use KM4 IPC to read/write secure physical OTP fuses
Previously all writing of the fuses happend through the KM4 using an IPC
mechamism. This was recently changed and the Linux driver can directly
write the fuses. However the access for secure fuses in the range
0x200 to 0x400 is only allowed from the secure world, so we need to
utilize the "realtek-otp-core" driver which will go through the KM4 to
read/write these regions as it was the case before.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>