rtkdrivers: otp: correctly map response buffer from physical memory
The OTP core/device will write the result to the physical address 0x6000bd50
in RAM.

When the kernel memory starts at 0x60000000 then this region will be covered
by the default mapping and phys_to_virt() works.

However in cases the kernel is not allowed to access this physical memory,
for example only allowing the kernel to access 0x61000000 and above, then
phys_to_virt() will not work correctly. So in that case we need to create
a temporary mapping when reading the response from the OTP.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>