drivers: rtkdrivers: adc: realtek-adc: don't disable digital input for non ADC pins
The Realtek ADC has 10 channels. Channels 0-5 correspond to Pins
PA0-PA5, and the remaining 4 channels are internal channels.

The pins PA0-PA8 can be chosen to be either digital or analog
input. This is done by setting the corresponding bit in the
digital/analog path selection register (0x420082F8).

The code was making the erroneous assumption that the channels above 5
mapped to the pins above PA5. This meant that enabling channel 6 (the
VBAT channel) in the DTS would disable digital input for PA6. This
caused PA6 to always read high when set as a GPIO.

Instead, only modify the digital/analog path selection register for
channels less or equal to 5.

Signed-off-by: Malik Tuwebti <malik.tuwebti@streamunlimited.com>