ASoC: meson: toddr: make S/PDIF in bit perfect [STREAMEIGHT-6756]
On the old kernel Amlogic used type 4 and endian 5 in case of 32 bit audio.
After experimenting it seems to add an extra padding after the first byte.
This is not what we expect to happen, instead we want to see the data as is
coming from the S/PDIF source.

So previously we had:
byte0 0x00 byte1 byte2
Now we have:
byte0 byte1 byte2 0x00

24 bit needs type = 4 while 32 bit needs type 3.
Verified through brute force experimentation.

Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>