S1832Config: enable CONFIG_PSTORE_RAM
This will allow us to check the kernel log, including the panic after
a crash of the kernel.

Per default the kernel enabled `CONFIG_PSTORE_DEFLATE_COMPRESS`, but it
does not seem to actually store the dump compressed in RAM as it can
be dumped inside the U-Boot as well, so rather just disabled it
explicitly.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>