S1832DTS: reserve 128 KiB of memory for ramoops
Put the region at an offset of 136 MiB (0x8800000) after the start of
the RAM. On other platforms we configure the reserved region to 128 MiB,
however on the A113 it seems like the beginning of various chunks of RAM
are overwritten by (probably) the memory training or something. So instead
start the reserved memory at 136 MiB which was experimentally determined
to work.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>