drivers: rtkdrivers: otp: fix logical fuse writing
In order to save fuses, writing a logical fuse should only write a
physical fuse if the value has changed. To do this, the OTP driver:
* Divides the logical address space into 16 byte sections.
* Retrieves the section containing the offset to be written.
* Determines the bytes/words that would be changed by the write.
* Only changes these bytes/words.

Instead of retrieving the full 16 byte section, code was previously only
retrieving 'bytes' bytes for the section. This meant the remaining
'16 - bytes' bytes of the section were filled with garbage. This caused
the driver to infer that bytes/words had been changed even when they
hadn't, and rewrite them unnecessarily - leading to excessive logical
fuse consumption.

Instead, retrieve 16 = RTK_OTP_LPGPKT_SIZE bytes for the section.

Signed-off-by: Malik Tuwebti <malik.tuwebti@streamunlimited.com>