ASoC: enforce chan min/max [STREAMEIGHT-6907]
The reason for not enforcing chan_min and chan_max when multiple CODEC DAIs
are connected to a single CPU DAI is that, the codecs can be connected
each to different channel, allowing producing multichannel output.

We generally do not support multichannel on Stream810 and presenting 32
channels towards userspace, was causing confusion, therefore rather limit
it to the least capable codec connected.

Signed-off-by: Radek Dostál <radek.dostal@streamunlimited.com>