Revert "optee: fix tee out of memory failure seen during kexec reboot" [STREAMEIGHT-6985]
This reverts commit f25889f93184db8b07a543cc2bbbb9a8fcaf4333.

When Cast starts after boot, it does some heavy op-tee operation through
the supplicant and a freeze can occur if the reboot is triggered during
this time. This can be fairly easily reproduced if the reboot is
called few seconds after the network is established.

First a context is open in tee_open():
open: ctx ffff00000edaf100, path: /dev/tee0
Then this context is never closed on shutdown and the whole system is stuck
inside __optee_disable_shm_cache() because optee_smc_do_call_with_arg() is
running indefinitely, and the optee_cq_wait_for_completion() never triggers.

Strangely we have not yet encountered this issue yet on imx8mm, which also uses
6.1 kernel. But we have different optee version there which might explain
the situation. We should consider dropping this revert if we upgrade optee
on a113 as well, or at least retest the reboot.

I tried closing the context, but the operation kept running inside optee
blocking the reboot. I have also tried upgrading optee-client, which also
did not help.

I have also tested the safeguard [0], and while it does improve the situation
it does not fully solve it. Sometimes kernel is stuck in reboot for 10-20
seconds then it finally reboots, but sometimes it is still fully stuck.

Reverting solves the problem fully, however this causes the cleanup to
not happen. That is not and issues because kexec is not a usecase for us,
so this workaround is suitable.

[0] https://lists.trustedfirmware.org/archives/list/op-tee@lists.trustedfirmware.org/thread/CDNLEZ7IFVRWTNKFV7JNMY63743XVSUT/

Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>
[rebase for 6.6]
Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>