rtkdrivers: otp_ipc: increase timeout for communicating with km4
The less logical fuses you have remaining, the longer it takes to write
to the fuses. If you don't have many logical fuses left, this can cause
a timeout on the kernel side, which in turn causes EINVAL to be passed
to user space.

Increase the timeout to 250 ms, so that we don't return a spurious error
to user space when there aren't many logical fuses remaining.

Signed-off-by: Malik Tuwebti <malik.tuwebti@streamunlimited.com>
[adjust commit message to include the total timeout time]
[also increase the timeout to for sending to the same value]
Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>