rtkdrivers: rtk-cpufreq: always configure the APLL during initialization
Previously this was called by chances because the `clock-frequency`
device-entry did not match any of the frequencies in the `freq_tab` list.

However the Realtek changed the behavior to unly run with a fixed APLL
frequency and use integer dividers, but to have the correct frequency
we need to configure the APLL once to the maximum frequency.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>