S195xDTS: enable crypto
This was disabled initially, but enabling it speeds up random number engine
initialization which makes the system boot much faster.

Crypto itself is defined in imx8mm.dtsi. There are 21 freescale device tree
files which do not disable it. Only one which disable it is
imx8mm-ddr4-evk-inmate.dts and that one seems to be for "jailhouse dual
Linux", where it is actually logical, that two Linux instances should not
be accessing same component. Therefore it should be safe to enable it.

Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>
[added information how it is used by NXP]
Signed-off-by: Radek Dostál <radek.dostal@streamunlimited.com>