rtkdrivers: cfg80211_fullmac: fix compatible of the ocp_node
Realtek forgot to rename this node here as well. It was not noticed by
them, as the code will early return and the only side-effect is that the
`axi_mem_start`, `axi_mem_end`, fields will not be populated in the
`paxi_data_global` global struct.

The only Realtek feature that relies on this is quering the TSF value
directly by using:
  cat /proc/1960/net/rtl8730e/wlan0/bcn_time

Doing this will crash the kernel as it relies on directly reading a
value at address `paxi_data_global->axi_mem_start + 0x560`.

In addition to that we have patched the `platform_device_init()` to map
the shared KM4 region in `km4_mem_end` and `km4_mem_start` so for us we
need the function to finish and also populate these fields as well.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>