HACK: drivers: soc: imx: simple-busfreq-imx8mm: delay entering low freq mode [STREAMEIGHT-7122]
Seems like the CAAM dies if the lowfreq mode is entered immediately.
Similar sleep is also implemented in the original busfreq driver:
drivers/soc/imx/busfreq-imx8mq.c:241

Adding the sleep in the busfreq driver has the benefit of being able to
detect if the system really needs to change freq. Doing it in the CAAM would
mean waiting every time an RNG is needed even when the freq is not changed.

Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>