drivers: rtkdrivers: misc: add sue-hsmp-realtek-mcu-ipc [STREAMEIGHT-6930]
This is the base driver that uses the Realtek IPC mechanism to
communicate with the M0 core using the HSMP protocol.

It implements the register access commands (read, write and update bits)
as a regmap bus and registers the regmap. Child drivers/devices (as
specified in the device-tree) can make use of this regmap by calling
`dev_get_regmap(pdev->dev.parent, NULL)` to get clean access to the
registers using the HSMP protocol over IPC.

This regmap bus approach allows us to have alternative transports like
I2C or UART in the future. As long as they provide us with a regmap,
the child drivers will not need any modifications.

Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>
[performed heavy cleanup and refactor]
[squashed multiple commits]
[added detailed commit message]
Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>
[remove extra brackets]
Signed-off-by: Radek Dostál <radek.dostal@streamunlimited.com>