drivers: misc: add sue-hsmp-state-controller [STREAMEIGHT-6930]
This driver uses a HSMP regmap to control the state and RGB led of a
host controller implementing the HSMP protocol.

The current state can be read and set via a sysfs file. In addition to
that the driver registers the RGB LEDs that can be used though the sysfs
as well.

NOTE: The RGB LEDs are only active in the `STATE_STREAMASDK` state, in all
      other states the external controller takes over the RGB LED control.

Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>
[performed minor cleanup and refactor]
[squashed multiple commits]
[added detailed commit message]
Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>
[remove extra tabs]
Signed-off-by: Radek Dostál <radek.dostal@streamunlimited.com>
Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>