ASoC: meson: switch pinmux when driver allowed to suspend and no stream is running [STREAMEIGHT-6344]
This allows us to pull all tdm lines low when there is no clk running.
This happens when "ignore_suspend == 0" and there is no stream running.

NOTE: The new pinmux entry must be called "suspend", if it is called
"sleep" the "soc-pcm" core will automatically select the "sleep" pinmux
when the PCM device is closed, which is not what we want.

Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>
[remove {} from one line statement]
Signed-off-by: Radek Dostál <radek.dostal@streamunlimited.com>
[rebase for 6.18]
Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>