drivers: firmware: arm_scpi: add optional SARADC dependency
On AXG platform the BL30 needs to be initialized to be able to properly read
the thermal sensor. If it is not initialized the BL30 will keep the lock and
the SARADC driver will fail to probe with a timeout so it cannot init the
ADC.

On 6.1 kernel the init order of these two drivers were masking this problem
but seems like on 6.6 the order is in favor of loading the SCPI first.

To work around this, check if the SARADC driver loaded while probing the SCPI
(by setting the platform drvdata) and defer until it has not.

Technically this depends on the SARADC to call the platform_set_drvdata()
last in the probe, but that seems to be the case for now.

Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>