rtkdrivers: thermal: use critical thermal point for the HW protection [STREAMEIGHT-6213]
This changes takes the configured "critical" thermal point and correctly
sets the `TM_BIT_HIGHCMP_PT_EN` bit. If this bit is set the module will
automatically reset (and keep in reset) the module as long as the temperature
is above this threshold.

Also use the passive trip point to configure the `high` (not `critical`)
trip point and interrupt.

The low threshold (as in a very low temperature) value and interrupt behavior
is kept unchanged.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>