ASoC: realtek: sport: implement `sue,continuous-clock` feature
This is similar to our other platforms, setting this flag does two
things:
1. Enables 48 kHz, 2-ch, 32-bit clocks (MCLK, BCLK, LRCLK) already
   during boot. Some audio codecs need at least the MCLK before you can
   configure them via I2C, so this will be achieved by that.
2. Keep the Tx clocks running after the playback has stopped. This will
   continue running the clocks at the previously played rate/setting.
   Especially on this platform this is needed as otherwise the I2S
   serializer might not finish the audio frame properly when stopping
   and we've observed issues with certain DACs here in the past.

The BCLK will be 64 * 48000 and the MCLK will be set according to the
`rtk,sport-mclk-multiplier` and `rtk,sport-mclk-fixed-max` properties.

On previous platforms this device-tree flag was just called
`continuous-clocks` without the `sue,` prefix. For quite some time we
have already used our prefix to indicate that a feature is
developed/customized by us, so we will also now use this flag with our
prefix even if it is not consistent with our other platforms.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>