rtkdrivers: cfg80211_fullmac: correctly handle phys offset in km4_phys_to_virt()
Previously the physical offset was hard-coded to 0x60000020, but we have
seen this change in the paste. It was correctly updated by Realtek in the
device-tree and mapped correclty, but this value also needed to be updated
in the code.

With this change we keep both track of the physical base and the virtual
based and can apply the offsets correctly.

Also removed useless `km4_mem_end` from `struc axi_data`.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>