drivers: rtkdrivers: drm: correctly resume when a suspend was aborted [STREAMEIGHT-6929]
When we try to enter a suspend but some other device fail to enter it, the
whole sequence is aborted, however in this case the `.resume_early`
callback will not be called and the device will be in the wrong state.

The `.resume` callback will be called correctly both in the successful and
also unsuccessful case to restore the device correctly.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>