sound: soc: axg-fifo: increase reported delay precision
When DMA engine consumes some data, axg_fifo_pcm_irq_block_thread
is invoked. In this handler snd_pcm_period_elapsed is called to notify
snd layer that some data was consumed. The snd stores the current time
and calls axg_fifo_pcm_pointer function to get a new position.
The axg_fifo_pcm_pointer is called when delay is requested from
the snd too. When the function returns a different position, current
timestamp is stored for this position.

The way the axg-fifo was implemented allows to call a snd_pcm_delay
to get the current delay without a need of checking when the delay was
changed by snd_pcm_period_elapsed. The issue is, the resolution of
the "FIFO_STATUS2" register is only 16 frames. The 16 frames is not a big
deal for multiroom synchronization as its less then 0.5ms for 48KHz.
The problem is, when we try to match speed of input with external clocks
with output running on internal clocks as number of processed data will
be rounded to 16 frames.

New implementation stores number of processed frames right after the
block is processed and snd stores current time in that moment. This way
we can get delay together with a time and calculate delay more precisely.
The same logic is used also on the master-stream195x.

Additionally, a regmap_read(fifo->map, FIFO_STATUS2, &fifo->dma_pos) call
was added in the __dma_enable function to ensure that the system starts
with the correct initial DMA position when it is enabled. Without this fix,
the system would assume that the position is 0, which is not always the
case. This could lead to sudden jumps in reported position (after it is
"fixed" by reading the actual value), which would result in discontinuity
detected by userspace applications.

Signed-off-by: Martin Geier <martin.geier@streamunlimited.com>
[extended commit message]
[squashed with "Report BATCH flag as driver now reports amount of
 consumed data in period-sized chunks."]
Signed-off-by: Radek Dostál <radek.dostal@streamunlimited.com>