ASoC: meson: check requested rate compatibility with TDM rate in SPDIF out [STREAMEIGHT-7551]
SPDIF can't set the parent clk because commit d0cd10e ("clk: meson:
axg-audio: enable SET_RATE_PARENT only for mst_mclk") only allows HIFI_PLL
to be set by TDM interface. If the TDM configures the PLL for a base rate
that is not compatible with the requested rate for the SPDIF playback the
rates can be mismatched. This results in CLK drift over time.

We can't allow the SPDIF to set the parent CLK because that would ruin the
settings set by the TDM interface.

With the transition to Pipewire it is not enough to issue a warning when
the rates are mismatched. The SPDIF has to refuse hw_params which are
not compatible with the parent PLL.
However checking the parent PLL would be rather cumbersome because:
1) it would require a walk up the clk tree or specifying the root in the DTS
2) The drift compensator can set it within ±500ppm which complicates maths

The cleanest way is to store the TDM rate when it is set, and calculate the
greatest common divisor with the requested SPDIF rate. If this GCD is greater
then 8000 we can be sure that the rates are compatible enough with each
other and the parent PLL can be divided cleanly into the requested rate.

Fixes: b78a7d4331a48 ("ASoC: meson: warn about SPDIF rate mismatch [STREAMEIGHT-6411]")

Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>
[rebase for 6.18]
Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>