ASoC: realtek: machine: add support for specifying codecs in device-tree
By default the AIF3 and AIF4 links use the dummy codec. This patch adds
support for overriding those with real codecs via the `sue,audio-codec`
device-tree property.

If the property is present, the driver will parse its entries and replace
the dummy codecs in AIF3 and AIF4 with the specified real codecs.

Example usage:

  // Specify both codecs
  sue,audio-codec = <&pcm9211 0>, <&pcm9211 2>;

  // Specify only the second codec, leave the first as dummy
  sue,audio-codec = <0>, <&pcm9211 2>;

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>