ASoC: realtek: machine: rework mute GPIOs and regulator handling
This makes the AMP and HP mute GPIOs optional as well as the regulator so
that the driver does not fail if they are not specified.

Also checking if the regulator is disabled/enabled before enabling/disabling
it is not correct and breaks the `enable_count` mechanism in case other
drivers might make use of the same regulator.

Apparently it might be that there is an unbalanced amount of
`SND_SOC_DAPM_PRE_PMU` and `SND_SOC_DAPM_PRE_PMD` calls, for that case we
manually keep track if we enabled/disabled the regulator in our own driver.

Signed-off-by: Martin Pietryka <martin.pietryka@streamunlimited.com>