S195xCONFIG: change modules to built-in drivers [STREAMEIGHT-7597]
The crypto modules can be built in, they don't take too much space on the
fit partition (around 300Kb). Although I think we don't use the majority
of them as they likely come from NXP defconfig, we are based on. The plan
is to keep them for now and clean them eventually as part of bigger
cleanup, which will make sure, that we enable same/similar crypto on all
our platforms.

We don't need I2C_HID_OF and RPCSEC_GSS_KRB5 so remove those.

Signed-off-by: Peter Suti <peter.suti@streamunlimited.com>
[polished commit message]
Signed-off-by: Radek Dostál <radek.dostal@streamunlimited.com>